requirements.txt
python_docx==1.1.2
setuptools==75.1.0
xlrd==2.0.1
xlwt==1.3.0
